Improve PEP8 compliance
Attempted to clean up fit_data.py by shortening some of the longer lines.
I intend to begin playing with some of the models again.
This may not be the best place for it, but I should mention that I would
like to make the entire PyPIMM utility a bit more modular, i.e. break it
up into smaller subprograms rather than have it be one monolithic program
like it is now. That way, if one step of the analysis is good and another
is not, I can just repeat that part.